requirements.txt
torchdiffeq~=0.2.5
alibi[tensorflow]~=0.9.6
numpy~=1.26.4
scipy~=1.14.1
cvxpy
scikit-learn~=1.6.0
neptune~=1.13.0
hydra-core
mlflow
umap-learn
nflows @ git+https://github.com/pfilo8/nflows.git
UMNN
pandas~=2.2.3
tqdm~=4.67.1
matplotlib
plotly
ipykernel
scikit-learn
mlflow
ruff

wandb~=0.19.1
torcheval~=0.0.7
openml~=0.15.0
omegaconf~=2.3.0
setuptools~=75.8.0
tensorflow~=2.14.1
scikit-image~=0.22.0